+++ src/js/leaflet.storage.forms.js
@@ -408,7 +408,8 @@ L.S.ElementHelper.PopupTemplate = L.S.ElementHelper.SelectAbstract.extend({
         ['Table', L._('Table')],
         ['GeoRSSImage', L._('GeoRSS (title + image)')],
         ['GeoRSSLink', L._('GeoRSS (only link)')],
-        ['SimplePanel', L._('Side panel')]
+        ['SimplePanel', L._('Side panel')],
+        ['Custom', L._('Custom template')]
     ],
 
     toJS: function () {
@@ -779,6 +780,7 @@ L.Storage.FormBuilder = L.Class.extend({
         iconClass: {handler: 'IconClassSwitcher', label: L._('type of icon')},
         iconUrl: {handler: 'IconUrl', label: L._('symbol of the icon')},
         popupTemplate: {handler: 'PopupTemplate', label: L._('template to use for the popup')},
+        popupCustom: {label: L._('Custom template'), handler: 'Textarea', helpEntries: ['dynamicProperties', 'textFormatting'], helpText: L._('You can use formatting and {properties} from your features.')},
         datalayer: {handler: 'DataLayerSwitcher', label: L._('Choose the layer of the feature')},
         moreControl: {handler: 'CheckBox', helpText: L._('Do you want to display the «more» control?')},
         datalayersControl: {handler: 'CheckBox', helpText: L._('Do you want to display the data layers control?')},